+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This version includes all revisions from r90 to rXX.
 
 ### Added
 - Current state of the game can now be saved.
+- Added mouse hover coordinates to debug.
+- Added Structure information to debug.
 
 ### Changed
 - Pulled old UI data files.